==> src/1_format_and_process/0 install_R_packages.R <==
install.packages("installr")
library(installr)
updateR()

install.packages("extrafont")
library(extrafont)
font_import()
loadfonts()

source("https://bioconductor.org/biocLite.R")
biocLite("SNPRelate")

install.packages("tidyverse")
install.packages("GGally")
install.packages("ggrepel")
install.packages("ape")
install.packages("plyr")
install.packages("adegenet")
install.packages("mmod")
install.packages("poppr")
install.packages("dbscan")
install.packages("scrime")
install.packages("circlize")
install.packages("dendextend")
install.packages("RColorBrewer")
# install.packages("stringr")
