==> src/1_format_and_process/1_generate_filtered_maps.R <==
library(tidyverse)

# set the base directories
base_in <- file.path("data", "marker_maps")
base_out <- file.path("data", "R", "marker_maps")

# load the pozniak genetic map
poz_gen_map <- read_csv(file.path(base_in, "pozniak_gen_map.csv"),
  na = "#N/A",
  col_names = c("marker", "group", "pos"),
  col_types = list(
    col_character(), col_factor(NULL), col_double()
  )
) %>%
  na.omit() %>%
  arrange(group, pos)

# load the wang genetic map
wang_gen_map <- read_csv(file.path(base_in, "wang_gen_map.csv"),
  na = "#N/A",
  col_names = c("marker", "group", "pos"),
  col_types = list(
      col_character(), col_factor(NULL), col_double()
  )
) %>%
  arrange(group, pos)

# filter the genetic maps so only markers present in the poz gen map are kept
# with those having a different position in the wang genetic map removed
poz_semi <- semi_join(poz_gen_map, wang_gen_map, by = c("marker", "group"))
poz_anti <- anti_join(poz_gen_map, wang_gen_map, by = c("marker", "group")) %>%
  anti_join(wang_gen_map, by = "marker")
poz_filtered <- poz_semi %>%
  bind_rows(poz_anti) %>%
  arrange(marker, group, pos)

# write the filtered genetic map out
ifelse(
  ! dir.exists(file.path(base_out)), dir.create(file.path(base_out)), FALSE
)
write_rds(poz_filtered, file.path(base_out, "pozniak_filtered_map.rds"))
